+++ internal/infrastructure/database/entschema/learned_word.go
@@ -45,10 +45,10 @@ func (LearnedWord) Fields() []ent.Field {
 		// FUTURE: Add lexeme_overrides JSONB field for per-lexeme mastery tracking
 		// This will enable advanced users to track mastery for specific word senses.
 		// Example structure: {"L123456": {"mastery": {...}, "note": "..."}}
-		field.Int16("mastery_listen").Default(0),
-		field.Int16("mastery_read").Default(0),
-		field.Int16("mastery_spell").Default(0),
-		field.Int16("mastery_pronounce").Default(0),
+		field.Int32("mastery_listen").Default(0),
+		field.Int32("mastery_read").Default(0),
+		field.Int32("mastery_spell").Default(0),
+		field.Int32("mastery_pronounce").Default(0),
 		field.Int32("mastery_overall").Default(0),
 		field.Time("review_last_review_at").Optional().Nillable(),
 		field.Time("review_next_review_at").Optional().Nillable(),
